initial commit
qtable and optimized lineup are both updated as we iterate through the lineup. This causes the optimized lineup to only have players from the end of the draft. This will need to be addressed in future commits so that either the optimized lineup is created after each round of optimization or we only select players that have the best values in the q table